requirements.txt
Django==2.1.7
djangorestframework==3.9.1
entrypoints==0.3
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.0
pytz==2018.9